Weather_refactor: #74 - controller refactoring
